+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/user_services_functions/start_user_services.go
@@ -3,13 +3,11 @@ package user_service_functions
 import (
 	"context"
 	"github.com/docker/go-connections/nat"
-	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/logs_collector_functions"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/shared_helpers"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/object_attributes_provider"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/container_status"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/enclave"
-	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/logs_collector"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/database_accessors/enclave_db/free_ip_addr_tracker"
@@ -98,8 +96,6 @@ func StartUserServices(
 	enclaveUuid enclave.EnclaveUUID,
 	services map[service.ServiceUUID]*service.ServiceConfig,
 	serviceRegistrations map[service.ServiceUUID]*service.ServiceRegistration,
-	logsCollector *logs_collector.LogsCollector,
-	logsCollectorAvailabilityChecker logs_collector_functions.LogsCollectorAvailabilityChecker,
 	objAttrsProvider object_attributes_provider.DockerObjectAttributesProvider,
 	freeIpProviderForEnclave *free_ip_addr_tracker.FreeIpAddrTracker,
 	dockerManager *docker_manager.DockerManager,
@@ -159,27 +155,6 @@ func StartUserServices(
 		return nil, nil, stacktrace.Propagate(err, "Couldn't get an object attribute provider for enclave '%v'", enclaveUuid)
 	}
 
-	// Check if the logs collector is available
-	// As the container logs are sent asynchronously we'd not know whether they're being received by the collector and there would be no errors if the collector never comes up
-	// The least we can do is check if the collector server is healthy before starting the user service, if in case it gets shut down later we can't do much about it anyway.
-	if err = logsCollectorAvailabilityChecker.WaitForAvailability(); err != nil {
-		return nil, nil,
-			stacktrace.Propagate(err, "An error occurred while waiting for the log container to become available")
-	}
-
-	//We use the public TCP address because the logging driver connection link is from the Docker demon to the logs collector container
-	//so the direction is from the host machine to the container inside the Docker cluster
-	logsCollectorServiceAddress, err := logsCollector.GetEnclaveNetworkTcpAddress()
-	if err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred getting the private TCP address")
-	}
-
-	//The following docker labels will be added into the logs stream which is necessary for creating new tags
-	//in the logs database and then use it for querying them to get the specific user service's logs
-	//even the 'enclaveUuid' value is used for Fluentbit to send it to Loki as the "X-Scope-OrgID" request's header
-	//due Loki is now configured to use multi tenancy, and we established this relation: enclaveUuid = tenantID
-	logsCollectorLabels := logs_collector_functions.GetKurtosisTrackedLogsCollectorLabels()
-
 	successfulStarts, failedStarts, err := runStartServiceOperationsInParallel(
 		ctx,
 		enclaveNetworkID,
@@ -188,8 +163,6 @@ func StartUserServices(
 		enclaveObjAttrsProvider,
 		freeIpProviderForEnclave,
 		dockerManager,
-		logsCollectorServiceAddress,
-		logsCollectorLabels,
 	)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred while trying to start services in parallel.")
@@ -221,8 +194,6 @@ func runStartServiceOperationsInParallel(
 	enclaveObjAttrsProvider object_attributes_provider.DockerEnclaveObjectAttributesProvider,
 	freeIpAddrProvider *free_ip_addr_tracker.FreeIpAddrTracker,
 	dockerManager *docker_manager.DockerManager,
-	logsCollectorAddress logs_collector.LogsCollectorAddress,
-	logsCollectorLabels logs_collector_functions.LogsCollectorLabels,
 ) (
 	map[service.ServiceUUID]*service.Service,
 	map[service.ServiceUUID]error,
@@ -247,8 +218,6 @@ func runStartServiceOperationsInParallel(
 			enclaveObjAttrsProvider,
 			freeIpAddrProvider,
 			dockerManager,
-			logsCollectorAddress,
-			logsCollectorLabels,
 		)
 	}
 
@@ -282,8 +251,6 @@ func createStartServiceOperation(
 	enclaveObjAttrsProvider object_attributes_provider.DockerEnclaveObjectAttributesProvider,
 	freeIpAddrProvider *free_ip_addr_tracker.FreeIpAddrTracker,
 	dockerManager *docker_manager.DockerManager,
-	logsCollectorAddress logs_collector.LogsCollectorAddress,
-	logsCollectorLabels logs_collector_functions.LogsCollectorLabels,
 ) operation_parallelizer.Operation {
 	id := serviceRegistration.GetName()
 	privateIpAddr := serviceRegistration.GetPrivateIP()
@@ -377,17 +344,6 @@ func createStartServiceOperation(
 			}
 		}
 
-		if logsCollectorAddress == "" {
-			return nil, stacktrace.NewError("Expected to have a logs collector server address value to send the user service logs, but it is empty")
-		}
-
-		logsCollectorAddressStr := string(logsCollectorAddress)
-		//The container will be configured to send the logs to the Fluentbit logs collector server
-		fluentdLoggingDriverCnfg := docker_manager.NewFluentdLoggingDriver(
-			logsCollectorAddressStr,
-			logsCollectorLabels,
-		)
-
 		createAndStartArgsBuilder := docker_manager.NewCreateAndStartContainerArgsBuilder(
 			containerImageName,
 			containerName.GetString(),
@@ -406,8 +362,6 @@ func createStartServiceOperation(
 			cpuAllocationMillicpus,
 		).WithMemoryAllocationMegabytes(
 			memoryAllocationMegabytes,
-		).WithLoggingDriver(
-			fluentdLoggingDriverCnfg,
 		).WithSkipAddingToBridgeNetworkIfStaticIpIsSet(
 			skipAddingUserServiceToBridgeNetwork,
 		)