+++ core/server/api_container/server/service_network/service_network.go
@@ -89,4 +89,6 @@ type ServiceNetwork interface {
 	UploadFilesArtifact(data []byte, artifactName string) (enclave_data_directory.FilesArtifactUUID, error)
 
 	IsNetworkPartitioningEnabled() bool
+
+	GetUniqueNameForFileArtifact() (string, error)
 }