+++ core/server/api_container/server/startosis_engine/startosis_packages/git_package_content_provider/git_package_content_provider_test.go
@@ -3,6 +3,8 @@ package git_package_content_provider
 import (
 	"fmt"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_constants"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
+	"github.com/kurtosis-tech/kurtosis/core/server/commons/yaml_parser"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/stretchr/testify/require"
 	"os"
@@ -41,7 +43,7 @@ func TestGitPackageProvider_SucceedsForValidPackageWithExplicitMasterSet(t *test
 
 	provider := NewGitPackageContentProvider(packageDir, packageTmpDir)
 
-	sampleStartosisModule := "github.com/kurtosis-tech/sample-startosis-load/sample.star@master"
+	sampleStartosisModule := "github.com/kurtosis-tech/sample-startosis-load/sample.star@main"
 	contents, err := provider.GetModuleContents(sampleStartosisModule)
 	require.Nil(t, err)
 	require.Equal(t, "a = \"World!\"\n", contents)
@@ -205,7 +207,7 @@ func Test_checkIfFileIsInAValidPackageInternal_somewhereInTheMiddle(t *testing.T
 	}
 
 	filePath := "/data/packages/root/subdir/subdir1/folder/some_file.txt"
-	actual, err := checkIfFileIsInAValidPackageInternal(filePath, "/data/packages", mockStatMethod)
+	actual, err := getKurtosisYamlPathForFileUrlInternal(filePath, "/data/packages", mockStatMethod)
 	require.Nil(t, err)
 	require.Equal(t, "/data/packages/root/subdir/subdir1/kurtosis.yml", actual)
 }
@@ -227,7 +229,7 @@ func Test_checkIfFileIsInAValidPackageInternal_packageIsSameAsWhereTheFileIs(t *
 	}
 
 	filePath := "/data/packages/root/subdir/some_file.txt"
-	actual, err := checkIfFileIsInAValidPackageInternal(filePath, "/data/packages", mockStatMethod)
+	actual, err := getKurtosisYamlPathForFileUrlInternal(filePath, "/data/packages", mockStatMethod)
 	require.Nil(t, err)
 	require.Equal(t, "/data/packages/root/subdir/kurtosis.yml", actual)
 }
@@ -249,7 +251,7 @@ func Test_checkIfFileIsInAValidPackageInternal_fileNotFound(t *testing.T) {
 	}
 
 	filePath := "/data/packages/root/subdir/some_file.txt"
-	actual, err := checkIfFileIsInAValidPackageInternal(filePath, "/data/packages", mockStatMethod)
+	actual, err := getKurtosisYamlPathForFileUrlInternal(filePath, "/data/packages", mockStatMethod)
 	require.Nil(t, err)
 	require.Equal(t, filePathToKurtosisYamlNotFound, actual)
 }
@@ -270,7 +272,7 @@ func Test_checkIfFileIsInAValidPackageInternal_unknownErrorOccurred(t *testing.T
 	}
 
 	filePath := "/data/packages/root/subdir/some_file.txt"
-	_, err := checkIfFileIsInAValidPackageInternal(filePath, "/data/packages", mockStatMethod)
+	_, err := getKurtosisYamlPathForFileUrlInternal(filePath, "/data/packages", mockStatMethod)
 	require.NotNil(t, err)
 	require.ErrorContains(t, err, fmt.Sprintf("An error occurred while locating %v in the path of '%v'", startosis_constants.KurtosisYamlName, filePath))
 }
@@ -281,7 +283,86 @@ func Test_checkIfFileIsInAValidPackageInternal_prefixMismatchError(t *testing.T)
 	}
 
 	filePath := "/packages/root/subdir/some_file.txt"
-	_, err := checkIfFileIsInAValidPackageInternal(filePath, "/data/packages", mockStatMethod)
+	_, err := getKurtosisYamlPathForFileUrlInternal(filePath, "/data/packages", mockStatMethod)
 	require.NotNil(t, err)
 	require.EqualError(t, err, fmt.Sprintf("Absolute path to file: %v must start with following prefix %v", filePath, "/data/packages"))
+}
+
+func Test_validatePackageNameMatchesKurtosisYamlLocation(t *testing.T) {
+	type args struct {
+		kurtosisYaml                    *yaml_parser.KurtosisYaml
+		absPathToPackageWithKurtosisYml string
+		packagesDir                     string
+	}
+	tests := []struct {
+		name string
+		args args
+		want *startosis_errors.InterpretationError
+	}{
+		{
+			name: "failure - mismatch package name and path (incorrect package name)",
+			args: args{
+				kurtosisYaml: &yaml_parser.KurtosisYaml{
+					PackageName: "github.com/author/repo/packageIncorrect",
+				},
+				absPathToPackageWithKurtosisYml: "/root/folder/author/repo/package/kurtosis.yml",
+				packagesDir:                     "/root/folder",
+			},
+			want: startosis_errors.NewInterpretationError("The package name in %v must match the location it is in. Package name is '%v' and kurtosis.yml is found here: '%v'", startosis_constants.KurtosisYamlName, "github.com/author/repo/packageIncorrect", "github.com/author/repo/package"),
+		},
+		{
+			name: "failure - mismatch package name and path (different location)",
+			args: args{
+				kurtosisYaml: &yaml_parser.KurtosisYaml{
+					PackageName: "github.com/author/repo",
+				},
+				absPathToPackageWithKurtosisYml: "/root/folder/author/repo/subfolder/kurtosis.yml",
+				packagesDir:                     "/root/folder",
+			},
+			want: startosis_errors.NewInterpretationError("The package name in %v must match the location it is in. Package name is '%v' and kurtosis.yml is found here: '%v'", startosis_constants.KurtosisYamlName, "github.com/author/repo", "github.com/author/repo/subfolder"),
+		},
+		{
+			name: "failure - contains a trailing '/'",
+			args: args{
+				kurtosisYaml: &yaml_parser.KurtosisYaml{
+					PackageName: "github.com/author/repo/subfolder/",
+				},
+				absPathToPackageWithKurtosisYml: "/root/folder/author/repo/subfolder/kurtosis.yml",
+				packagesDir:                     "/root/folder",
+			},
+			want: startosis_errors.NewInterpretationError("Kurtosis package name cannot have trailing \"/\"; package name: %v and kurtosis.yml is found at: %v", "github.com/author/repo/subfolder/", "github.com/author/repo/subfolder/kurtosis.yml"),
+		},
+		{
+			name: "success - kurtosis.yml found in repo folder",
+			args: args{
+				kurtosisYaml: &yaml_parser.KurtosisYaml{
+					PackageName: "github.com/author/repo",
+				},
+				absPathToPackageWithKurtosisYml: "/root/folder/author/repo/kurtosis.yml",
+				packagesDir:                     "/root/folder",
+			},
+			want: nil,
+		},
+		{
+			name: "success - kurtosis.yml found in sub folder folder",
+			args: args{
+				kurtosisYaml: &yaml_parser.KurtosisYaml{
+					PackageName: "github.com/author/repo/subfolder",
+				},
+				absPathToPackageWithKurtosisYml: "/root/folder/author/repo/subfolder/kurtosis.yml",
+				packagesDir:                     "/root/folder",
+			},
+			want: nil,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := validatePackageNameMatchesKurtosisYamlLocation(tt.args.kurtosisYaml, tt.args.absPathToPackageWithKurtosisYml, tt.args.packagesDir)
+			if tt.want == nil {
+				require.Nil(t, err)
+			} else {
+				require.EqualError(t, err, tt.want.Error())
+			}
+		})
+	}
 }